refactor(i18n): use vocabularies i18next for funding ui
