+++ .tx/config
@@ -7,7 +7,6 @@
 # modify it under the terms of the MIT License; see LICENSE file for more
 # details.
 
-# TODO: Transifex integration
 #
 # 1) Create message catalog:
 #    $ python setup.py extract_messages
@@ -30,5 +29,38 @@ host = https://app.transifex.com
 [o:inveniosoftware:p:invenio:r:invenio-vocabularies-messages]
 file_filter = invenio_vocabularies/translations/<lang>/LC_MESSAGES/messages.po
 source_file = invenio_vocabularies/translations/messages.pot
+source_lang = en
+type = PO
+
+# Translate JavaScript strings
+# 1) Navigate to the invenio_vocabularies/assets/semantic-ui/translations/invenio_vocabularies folder
+# 2) Install i18n dev dependencies
+#    npm install
+# 3) Add a new language
+#    npm run init_catalog lang <lang>
+# 4) Extract translation keys/values
+#    $ npm run extract_messages
+# 5) Update the ./messages/index.js file
+#    import TRANSLATE_<lang> from './<lang>/translations.json'
+#    export const translations = {
+#      ...rest,
+#      <lang>: { translation: TRANSLATE_<lang> }
+#    }
+# 6) Install the transifex-client
+#    $ pip install transifex-client
+# 7) Push source (.pot) and translations (.po) to Transifex
+#    $ tx push -s -t
+# 8) Pull translations for a single language from Transifex
+#    $ tx pull -l <lang>
+# 9) Pull translations for all languages from Transifex
+#    $ tx pull -a
+# 10) Compile .po files for all languages
+#    $ npm run compile_catalog
+# 11) Convert .po file for a single language
+#    $ npm run compile_catalog lang <lang>
+
+[o:inveniosoftware:p:invenio:r:invenio-vocabularies-messages-ui]
+file_filter = invenio_vocabularies/assets/semantic-ui/translations/invenio_vocabularies/messages/<lang>/messages.po
+source_file = invenio_vocabularies/assets/semantic-ui/translations/invenio_vocabularies/translations.pot
 source_lang = en
 type = PO
