+++ invenio_vocabularies/assets/semantic-ui/js/invenio_vocabularies/src/contrib/forms/Funding/CustomAwardForm.js
@@ -9,7 +9,7 @@ import PropTypes from "prop-types";
 import React from "react";
 import { Form, Header } from "semantic-ui-react";
 import { TextField, RemoteSelectField } from "react-invenio-forms";
-import { i18next } from "@translations/invenio_rdm_records/i18next";
+import { i18next } from "@translations/invenio_vocabularies/i18next";
 import _isEmpty from "lodash/isEmpty";
 
 import Overridable from "react-overridable";
