+++ invenio_vocabularies/assets/semantic-ui/js/invenio_vocabularies/src/contrib/forms/Funding/FundingField.js
@@ -16,7 +16,7 @@ import { FieldLabel, FeedbackLabel } from "react-invenio-forms";
 import { FundingFieldItem } from "./FundingFieldItem";
 import FundingModal from "./FundingModal";
 
-import { i18next } from "@translations/invenio_rdm_records/i18next";
+import { i18next } from "@translations/invenio_vocabularies/i18next";
 
 import Overridable from "react-overridable";
