+++ invenio_vocabularies/assets/semantic-ui/js/invenio_vocabularies/src/contrib/forms/Funding/FundingModal.js
@@ -5,7 +5,7 @@
 // Invenio is free software; you can redistribute it and/or modify it
 // under the terms of the MIT License; see LICENSE file for more details.
 
-import { i18next } from "@translations/invenio_rdm_records/i18next";
+import { i18next } from "@translations/invenio_vocabularies/i18next";
 import { Formik } from "formik";
 import PropTypes from "prop-types";
 import React, { useState } from "react";
