+++ invenio_vocabularies/assets/semantic-ui/js/invenio_vocabularies/src/contrib/forms/Funding/NoAwardResults.js
@@ -8,7 +8,7 @@
 import PropTypes from "prop-types";
 import React from "react";
 import { Segment } from "semantic-ui-react";
-import { i18next } from "@translations/invenio_rdm_records/i18next";
+import { i18next } from "@translations/invenio_vocabularies/i18next";
 
 export function NoAwardResults({ switchToCustom }) {
   return (
