+++ invenio_vocabularies/assets/semantic-ui/translations/invenio_vocabularies/package.json
@@ -13,7 +13,6 @@
       "en",
       "es",
       "et",
-      "et_EE",
       "fa",
       "fr",
       "gl",
