+++ invenio_vocabularies/assets/semantic-ui/translations/invenio_vocabularies/scripts/compileCatalog.js
@@ -1,4 +1,4 @@
-// This file is part of React-Invenio-Deposit
+// This file is part of Invenio-Vocabularies
 //
 // Invenio-administration is free software; you can redistribute it and/or modify it
 // under the terms of the MIT License; see LICENSE file for more details.
